# server/openapi.yaml
openapi: 3.1.0
info:
  title: Finance Query v2
  version: 3.0.0
  description: High-performance financial data API server built on finance-query library
servers:
  - url: "https://finance-query.com"
  - url: "http://localhost:8000"
paths:
...
  "/v2/indicators/{symbol}":
    get:
      operationId: getIndicators
      summary: Get technical indicators
      parameters:
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: interval
          in: query
          required: false
          schema:
            description: "Candle interval (default: 1d)"
            $ref: "#/components/schemas/Interval"
        - name: range
          in: query
          required: false
          schema:
            description: "Time range (default: 1mo)"
            $ref: "#/components/schemas/TimeRange"
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlIndicatorsSummary"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  "/v2/quote-type/{symbol}":
    get:
      operationId: getQuoteType
      summary: Get quote type
      parameters:
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlQuoteTypeData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  "/v2/quote/{symbol}":
    get:
      operationId: getQuote
      summary: Get quote for a symbol
      parameters:
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: format
          in: query
          required: false
          schema:
            description: "Value format: raw, pretty, or both (default: raw)"
            $ref: "#/components/schemas/ValueFormat"
        - name: lang
          in: query
          required: false
          schema:
            description: "Target language for translated text fields (BCP 47, e.g. \"ja\", \"zh-Hant\");\nfalls back to the Accept-Language header"
            type: string
        - name: logo
          in: query
          required: false
          schema:
            description: "Whether to include company logo URL (default: false)"
            type: boolean
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlQuote"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  "/v2/risk/{symbol}":
    get:
      operationId: getRisk
      summary: Risk analytics
      parameters:
        - name: benchmark
          in: query
          required: false
          schema:
            description: "Optional benchmark symbol for beta calculation (e.g., \"SPY\")"
            type: string
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: interval
          in: query
          required: false
          schema:
            description: "Candle interval (default: 1d)"
            $ref: "#/components/schemas/Interval"
        - name: range
          in: query
          required: false
          schema:
            description: "Time range (default: 1mo)"
            $ref: "#/components/schemas/TimeRange"
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlRiskSummary"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      description: "Generic REST error body, emitted on 4xx/5xx by the GraphQL bridge and by\nhandlers rejecting invalid path/query input."
      type: object
      properties:
        error:
          description: Human-readable error message.
          type: string
        status:
          description: "HTTP status code, duplicated in the body."
          type: integer
          format: int32
          minimum: 0
      required:
        - error
        - status
...
    GqlAroonData:
      type: object
      properties:
        aroonDown:
          type: number
          format: double
        aroonUp:
          type: number
          format: double
...
    GqlBollingerBandsData:
      type: object
      properties:
        lower:
          type: number
          format: double
        middle:
          type: number
          format: double
        upper:
          type: number
          format: double
    GqlBullBearPowerData:
      type: object
      properties:
        bearPower:
          type: number
          format: double
        bullPower:
          type: number
          format: double
...
    GqlDonchianChannelsData:
      type: object
      properties:
        lower:
          type: number
          format: double
        middle:
          type: number
          format: double
        upper:
          type: number
          format: double
...
    GqlElderRayData:
      type: object
      properties:
        bearPower:
          type: number
          format: double
        bullPower:
          type: number
          format: double
...
    GqlFibonacciLevels:
      description: "Mirrors `finance_query::indicators::FibonacciLevels`, which has no serde\nrename of its own (plain snake_case keys)."
      type: object
      properties:
        level236:
          type: number
          format: double
        level382:
          type: number
          format: double
        level50:
          type: number
          format: double
        level618:
          type: number
          format: double
        level786:
          type: number
          format: double
        swingHigh:
          type: number
          format: double
        swingLow:
          type: number
          format: double
...
    GqlHeikinAshiCandle:
      description: "Mirrors `finance_query::Candle` (camelCase serde keys)."
      type: object
      properties:
        adjClose:
          type: number
          format: double
        close:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        open:
          type: number
          format: double
        timestamp:
          type: integer
          format: int64
        volume:
          type: integer
          format: int64
    GqlIchimokuData:
      type: object
      properties:
        baseLine:
          type: number
          format: double
        conversionLine:
          type: number
          format: double
        laggingSpan:
          type: number
          format: double
        leadingSpanA:
          type: number
          format: double
        leadingSpanB:
          type: number
          format: double
...
    GqlIndicatorsSummary:
      description: All technical indicators for a symbol (latest values).
      type: object
      properties:
        accumulationDistribution:
          type: number
          format: double
        adx14:
          type: number
          format: double
        alma9:
          type: number
          format: double
        aroon:
          $ref: "#/components/schemas/GqlAroonData"
        atr14:
          type: number
          format: double
        awesomeOscillator:
          type: number
          format: double
        balanceOfPower:
          type: number
          format: double
        bollingerBands:
          $ref: "#/components/schemas/GqlBollingerBandsData"
        bullBearPower:
          $ref: "#/components/schemas/GqlBullBearPowerData"
        cci20:
          type: number
          format: double
        chaikinOscillator:
          type: number
          format: double
        choppinessIndex14:
          type: number
          format: double
        cmf20:
          type: number
          format: double
        cmo14:
          type: number
          format: double
        coppockCurve:
          type: number
          format: double
        dema20:
          type: number
          format: double
        donchianChannels:
          $ref: "#/components/schemas/GqlDonchianChannelsData"
        elderRayIndex:
          $ref: "#/components/schemas/GqlElderRayData"
        ema10:
          type: number
          format: double
        ema100:
          type: number
          format: double
        ema20:
          type: number
          format: double
        ema200:
          type: number
          format: double
        ema50:
          type: number
          format: double
        fibonacciPivotPoints:
          $ref: "#/components/schemas/GqlPivotPointsData"
        fibonacciRetracement50:
          $ref: "#/components/schemas/GqlFibonacciLevels"
        heikinAshi:
          $ref: "#/components/schemas/GqlHeikinAshiCandle"
        hma20:
          type: number
          format: double
        ichimoku:
          $ref: "#/components/schemas/GqlIchimokuData"
        keltnerChannels:
          $ref: "#/components/schemas/GqlKeltnerChannelsData"
        macd:
          $ref: "#/components/schemas/GqlMacdData"
        mcginleyDynamic20:
          type: number
          format: double
        mfi14:
          type: number
          format: double
        momentum10:
          type: number
          format: double
        obv:
          type: number
          format: double
        parabolicSar:
          type: number
          format: double
        pivotPoints:
          $ref: "#/components/schemas/GqlPivotPointsData"
        roc12:
          type: number
          format: double
        rsi14:
          type: number
          format: double
        sma10:
          type: number
          format: double
        sma100:
          type: number
          format: double
        sma20:
          type: number
          format: double
        sma200:
          type: number
          format: double
        sma50:
          type: number
          format: double
        stochastic:
          $ref: "#/components/schemas/GqlStochasticData"
        stochasticRsi:
          $ref: "#/components/schemas/GqlStochasticData"
        supertrend:
          $ref: "#/components/schemas/GqlSuperTrendData"
        tema20:
          type: number
          format: double
        trueRange:
          type: number
          format: double
        vwap:
          type: number
          format: double
        vwma20:
          type: number
          format: double
        williamsR14:
          type: number
          format: double
        wma10:
          type: number
          format: double
        wma100:
          type: number
          format: double
        wma20:
          type: number
          format: double
        wma200:
          type: number
          format: double
        wma50:
          type: number
          format: double
        zigzagLastPivot:
          $ref: "#/components/schemas/GqlZigZagPoint"
...
    GqlKeltnerChannelsData:
      type: object
      properties:
        lower:
          type: number
          format: double
        middle:
          type: number
          format: double
        upper:
          type: number
          format: double
...
    GqlMacdData:
      type: object
      properties:
        histogram:
          type: number
          format: double
        macd:
          type: number
          format: double
        signal:
          type: number
          format: double
...
    GqlPivotPointsData:
      type: object
      properties:
        pivot:
          type: number
          format: double
        r1:
          type: number
          format: double
        r2:
          type: number
          format: double
        r3:
          type: number
          format: double
        s1:
          type: number
          format: double
        s2:
          type: number
          format: double
        s3:
          type: number
          format: double
...
    GqlQuoteTypeData:
      description: "Quote type metadata for a symbol (exchange, timezone, identifiers)."
      type: object
      properties:
        exchange:
          type: string
        firstTradeDateEpochUtc:
          type: integer
          format: int64
        gmtOffSetMilliseconds:
          type: integer
          format: int64
        longName:
          type: string
        maxAge:
          type: integer
          format: int64
        messageBoardId:
          type: string
        quoteType:
          type: string
        shortName:
          type: string
        symbol:
          type: string
        timeZoneFullName:
          type: string
        timeZoneShortName:
          type: string
        underlyingSymbol:
          type: string
        uuid:
          type: string
...
    GqlRiskSummary:
      description: Risk/performance summary for a symbol.
      type: object
      properties:
        beta:
          type: number
          format: double
        calmar:
          type: number
          format: double
        cvar95:
          type: number
          format: double
        cvar99:
          type: number
          format: double
        informationRatio:
          type: number
          format: double
        kelly:
          type: number
          format: double
        maxDrawdown:
          type: number
          format: double
        maxDrawdownRecoveryPeriods:
          type: number
          format: double
        omega:
          type: number
          format: double
        parametricCvar95:
          type: number
          format: double
        parametricVar95:
          type: number
          format: double
        sharpe:
          type: number
          format: double
        sortino:
          type: number
          format: double
        trackingError:
          type: number
          format: double
        ulcerIndex:
          type: number
          format: double
        var95:
          type: number
          format: double
        var99:
          type: number
          format: double
      required:
        - var95
        - var99
        - parametricVar95
        - cvar95
        - cvar99
        - parametricCvar95
        - omega
        - kelly
        - maxDrawdown
        - ulcerIndex
...
    GqlStochasticData:
      description: "Mirrors `finance_query::indicators::summary::StochasticData`, whose `k`/`d`\nfields serialize as `\"%K\"`/`\"%D\"` (a custom per-field rename, not plain\ncamelCase) — must match exactly for deserialization to find them."
      type: object
      properties:
        d:
          type: number
          format: double
        k:
          type: number
          format: double
    GqlSuperTrendData:
      type: object
      properties:
        trend:
          type: string
        value:
          type: number
          format: double
...
    Interval:
      description: Chart intervals
      type: string
      enum:
        - 1m
        - 2m
        - 5m
        - 15m
        - 30m
        - 1h
        - 90m
        - 1d
        - 5d
        - 1wk
        - 1mo
        - 3mo
...
    RateLimitError:
      description: "Rate limit error body, returned with HTTP 429 and a `Retry-After` header."
      type: object
      properties:
        error:
          description: "Always \"Rate limit exceeded\"."
          type: string
        message:
          description: Human-readable message including the retry delay.
          type: string
        retryAfterSeconds:
          description: Seconds until the next request would be admitted.
          type: integer
          format: int64
          minimum: 0
      required:
        - error
        - message
        - retryAfterSeconds
...
    TimeRange:
      description: Time ranges for chart data
      type: string
      enum:
        - 1d
        - 5d
        - 1mo
        - 3mo
        - 6mo
        - 1y
        - 2y
        - 5y
        - 10y
        - ytd
        - max
...
    ValueFormat:
      description: Value format for API responses
      type: string
      enum:
        - raw
        - pretty
        - both
...